Ready!

+++ WorldOfEnglishWord/Controllers/Cards/CardEnFragment.cs
@@ -1,12 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-
-using Android.App;
-using Android.Content;
 using Android.OS;
-using Android.Runtime;
 using Android.Views;
 using Android.Widget;
 
@@ -70,7 +64,7 @@ namespace WorldOfEnglishWord.Controllers.Cards
                 return "-";
             }
 
-            float percent = (float)(statistic[0] / statistic[1]) * 100;
+            double percent = (double)statistic[0] / statistic[1] * 100;
 
             if (percent < 51)
             {
